+++ examples/react/getting-started/src/App.tsx
@@ -12,6 +12,7 @@ import {
   TrendingItems,
   Carousel,
   Chat,
+  ChatTrigger,
   FilterSuggestions,
   CurrentRefinements,
 } from 'react-instantsearch';
@@ -96,6 +97,7 @@ export function App() {
             feedback={true}
             itemComponent={ItemComponent}
           />
+          <ChatTrigger />
         </InstantSearch>
       </div>
     </div>