+++ packages/instantsearch.js/src/connectors/index.ts
@@ -57,4 +57,5 @@ export { default as connectFrequentlyBoughtTogether } from './frequently-bought-
 export { default as connectLookingSimilar } from './looking-similar/connectLookingSimilar';
 export { default as connectChat } from './chat/connectChat';
 export { default as connectFeeds } from './feeds/connectFeeds';
+export { default as connectChatTrigger } from './chat/connectChatTrigger';
 export { default as connectFilterSuggestions } from './filter-suggestions/connectFilterSuggestions';