+++ packages/vue-instantsearch/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.26.5](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.4...vue-instantsearch@4.26.5) (2026-05-26)
+
+**Note:** Version bump only for package vue-instantsearch
+
+
+
+
+
 ## [4.26.4](https://github.com/algolia/instantsearch/compare/vue-instantsearch@4.26.3...vue-instantsearch@4.26.4) (2026-05-19)
 
 **Note:** Version bump only for package vue-instantsearch